# public/swagger/swagger.json
{
    "basePath": "/",
...
    "host": "localhost:7465",
    "info": {
        "contact": {
            "email": "pat@cappelaere.com"
        },
        "description": "This is OJO Server REST Application Programming Interface Defintions.",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "termsOfService": "http://ojo-bot.herokuapp.com/terms/",
        "title": "OJO-BOT API",
        "version": "1.0.0"
    },
    "paths": {
        "/opensearch": {
            "get": {
                "description": "Generic Search for Actionable Product",
                "operationId": "opensearch",
                "parameters": [
                    {
                        "default": "precip_3hrs",
                        "description": "Product query",
                        "enum": [
                            "precip_30mn",
                            "precip_3hrs",
                            "precip_1d",
                            "precip_7d"
                        ],
                        "in": "query",
                        "name": "q",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "default": 0,
                        "description": "Latitude",
                        "in": "query",
                        "maximum": 90.0,
                        "minimum": -90.0,
                        "name": "lat",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "default": 0,
                        "description": "Longitude",
                        "in": "query",
                        "maximum": 180.0,
                        "minimum": -180.0,
                        "name": "lon",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "default": 2,
                        "description": "Limit number of responses",
                        "in": "query",
                        "maximum": 1000,
                        "minimum": 1,
                        "name": "limit",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "default": "gpm",
                        "description": "Product sources",
                        "enum": [
                            "gpm",
                            "trmm"
                        ],
                        "in": "query",
                        "name": "sources",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "description": "Start Time",
                        "format": "date",
                        "in": "query",
                        "name": "startTime",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "description": "End Time",
                        "format": "date",
                        "in": "query",
                        "name": "endTime",
                        "required": false,
                        "type": "string"
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "summary": "Generic OpenSearch",
                "tags": [
                    "opensearch"
                ]
            }
        },
        "/opensearch/precip_gpm_1d": {
            "get": {
                "description": "OpenSearch for Global 1-day Accumulated Rainfall (GPM)",
                "operationId": "opensearch",
                "parameters": [
                    {
                        "default": "precip_1d",
                        "description": "Product query",
                        "in": "query",
                        "name": "q",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "default": 0,
                        "description": "Latitude",
                        "in": "query",
                        "maximum": 90.0,
                        "minimum": -90.0,
                        "name": "lat",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "default": 0,
                        "description": "Longitude",
                        "in": "query",
                        "maximum": 180.0,
                        "minimum": -180.0,
                        "name": "lon",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "default": 10,
                        "description": "Limit number of responses",
                        "in": "query",
                        "maximum": 1000,
                        "minimum": 1,
                        "name": "limit",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "default": "gpm",
                        "description": "Product sources",
                        "in": "query",
                        "name": "sources",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "description": "Start Time",
                        "format": "date",
                        "in": "query",
                        "name": "startTime",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "description": "End Time",
                        "format": "date",
                        "in": "query",
                        "name": "endTime",
                        "required": false,
                        "type": "string"
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "summary": "OpenSearch for Global 1-day Accumulated Rainfall Product (GPM)",
                "tags": [
                    "opensearch"
                ]
            }
        },
        "/opensearch/precip_gpm_30mn": {
            "get": {
                "description": "OpenSearch for Golbal 30mn Accumulated Rainfall (GPM)",
                "operationId": "opensearch",
                "parameters": [
                    {
                        "default": "precip_30mn",
                        "description": "Product query",
                        "in": "query",
                        "name": "q",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "default": 0,
                        "description": "Latitude",
                        "in": "query",
                        "maximum": 90.0,
                        "minimum": -90.0,
                        "name": "lat",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "default": 0,
                        "description": "Longitude",
                        "in": "query",
                        "maximum": 180.0,
                        "minimum": -180.0,
                        "name": "lon",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "default": 10,
                        "description": "Limit number of responses",
                        "in": "query",
                        "maximum": 1000,
                        "minimum": 1,
                        "name": "limit",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "default": "gpm",
                        "description": "Product sources",
                        "in": "query",
                        "name": "sources",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "description": "Start Time",
                        "format": "date",
                        "in": "query",
                        "name": "startTime",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "description": "End Time",
                        "format": "date",
                        "in": "query",
                        "name": "endTime",
                        "required": false,
                        "type": "string"
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "summary": "OpenSearch for Global 30mn Accumulated Rainfall Product (GPM)",
                "tags": [
                    "opensearch"
                ]
            }
        },
        "/opensearch/precip_gpm_3hrs": {
            "get": {
                "description": "OpenSearch for Global 3hrs Accumulated Rainfall (GPM)",
                "operationId": "opensearch",
                "parameters": [
                    {
                        "default": "precip_3hrs",
                        "description": "Product query",
                        "in": "query",
                        "name": "q",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "default": 0,
                        "description": "Latitude",
                        "in": "query",
                        "maximum": 90.0,
                        "minimum": -90.0,
                        "name": "lat",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "default": 0,
                        "description": "Longitude",
                        "in": "query",
                        "maximum": 180.0,
                        "minimum": -180.0,
                        "name": "lon",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "default": 2,
                        "description": "Limit number of responses",
                        "in": "query",
                        "maximum": 1000,
                        "minimum": 1,
                        "name": "limit",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "default": "gpm",
                        "description": "Product sources",
                        "in": "query",
                        "name": "sources",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "description": "Start Time",
                        "format": "date",
                        "in": "query",
                        "name": "startTime",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "description": "End Time",
                        "format": "date",
                        "in": "query",
                        "name": "endTime",
                        "required": false,
                        "type": "string"
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "summary": "OpenSearch for Global 3Hr Accumulated Rainfall Product (GPM)",
                "tags": [
                    "opensearch"
                ]
            }
        },
        "/opensearch/precip_gpm_7d": {
            "get": {
                "description": "OpenSearch for Global 7-day Accumulated Rainfall (GPM)",
                "operationId": "opensearch",
                "parameters": [
                    {
                        "default": "precip_7d",
                        "description": "Product query",
                        "in": "query",
                        "name": "q",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "default": 0,
                        "description": "Latitude",
                        "in": "query",
                        "maximum": 90.0,
                        "minimum": -90.0,
                        "name": "lat",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "default": 0,
                        "description": "Longitude",
                        "in": "query",
                        "maximum": 180.0,
                        "minimum": -180.0,
                        "name": "lon",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "default": 10,
                        "description": "Limit number of responses",
                        "in": "query",
                        "maximum": 1000,
                        "minimum": 1,
                        "name": "limit",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "default": "gpm",
                        "description": "Product sources",
                        "in": "query",
                        "name": "sources",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "description": "Start Time",
                        "format": "date",
                        "in": "query",
                        "name": "startTime",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "description": "End Time",
                        "format": "date",
                        "in": "query",
                        "name": "endTime",
                        "required": false,
                        "type": "string"
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "summary": "OpenSearch for Global 7-day Accumulated Rainfall Product (GPM)",
                "tags": [
                    "opensearch"
                ]
            }
        },
        "/opensearch/precip_trmm_1d": {
            "get": {
                "description": "OpenSearch for Regional 1-day Accumulated Rainfall (TRMM)",
                "operationId": "opensearch",
                "parameters": [
                    {
                        "default": "precip_1d",
                        "description": "Product query",
                        "in": "query",
                        "name": "q",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "default": 9.84,
                        "description": "Latitude",
                        "in": "query",
                        "maximum": 90.0,
                        "minimum": -90.0,
                        "name": "lat",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "default": -84.14,
                        "description": "Longitude",
                        "in": "query",
                        "maximum": 180.0,
                        "minimum": -180.0,
                        "name": "lon",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "default": 10,
                        "description": "Limit number of responses",
                        "in": "query",
                        "maximum": 1000,
                        "minimum": 1,
                        "name": "limit",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "default": "trmm",
                        "description": "Product sources",
                        "in": "query",
                        "name": "sources",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "description": "Start Time",
                        "format": "date",
                        "in": "query",
                        "name": "startTime",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "description": "End Time",
                        "format": "date",
                        "in": "query",
                        "name": "endTime",
                        "required": false,
                        "type": "string"
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "summary": "OpenSearch for Regional 1-day Accumulated Rainfall Product (TRMM)",
                "tags": [
                    "opensearch"
                ]
            }
        },
...
    },
    "schemes": [
        "http"
    ],
    "securityDefinitions": {
        "api_key": {
            "in": "header",
            "name": "api_key",
            "type": "apiKey"
        },
        "petstore_auth": {
            "authorizationUrl": "http://petstore.swagger.io/api/oauth/dialog",
            "flow": "implicit",
            "scopes": {
                "read:pets": "read your pets",
                "write:pets": "modify pets in your account"
            },
            "type": "oauth2"
        }
    },
    "swagger": "2.0",
...
}